+++ m8-files/src/web.rs
@@ -94,6 +94,44 @@ pub fn show_phrase(song: &WasmSong, phrase: usize) -> String {
 }
 
 #[wasm_bindgen]
+pub fn rename_instrument(song: &mut WasmSong, instrument: usize, new_name: String) -> Result<bool, String> {
+    if instrument >= Song::N_INSTRUMENTS {
+        return Err(format!("Error invalid source instrument number {instrument}"));
+    }
+
+    if new_name.len() > 12 {
+        return Err(format!("name \"{new_name}\" is too long (12 max) "));
+    }
+
+    song.song.instruments[instrument].set_name(new_name);
+
+    Ok(true)
+}
+
+
+#[wasm_bindgen]
+pub fn renumber_instrument(song: &mut WasmSong, instrument: usize, to_instrument: usize) -> Result<bool, String> {
+    if instrument >= Song::N_INSTRUMENTS {
+        return Err(format!("Error invalid source instrument number {instrument}"));
+    }
+
+    if to_instrument >= Song::N_INSTRUMENTS {
+        return Err(format!("Error invalid source destination number {to_instrument}"));
+    }
+
+    if ! song.song.instruments[to_instrument].is_empty() {
+        return Err(format!("Destination instrument{to_instrument} is not empty"));
+    }
+
+    let mut remapper = Remapper::default();
+    remapper.instrument_mapping.mapping[instrument] = to_instrument as u8;
+    remapper.instrument_mapping.to_move.push(instrument as u8);
+    remapper.renumber(&mut song.song);
+
+    Ok(true)
+}
+
+#[wasm_bindgen]
 pub fn renumber_chain(song: &mut WasmSong, chain: usize, to_chain: usize) -> Result<bool, String> {
     if chain >= Song::N_CHAINS { return Err(format!("Error invalid source chain number {chain}")) }
     if to_chain >= Song::N_CHAINS { return Err(format!("Error invalid destination chain number {to_chain}")) }
@@ -105,6 +143,7 @@ pub fn renumber_chain(song: &mut WasmSong, chain: usize, to_chain: usize) -> Res
     }
 
     song.song.chains[to_chain] = song.song.chains[chain].clone();
+    song.song.chains[chain].clear();
 
     let song_steps = &mut song.song.song.steps;
     for i in 0 .. song_steps.len() {
@@ -113,7 +152,24 @@ pub fn renumber_chain(song: &mut WasmSong, chain: usize, to_chain: usize) -> Res
         }
     }
 
-    log(&format!("RENUMBERED {chain} -> {to_chain}"));
+    Ok(true)
+}
+
+#[wasm_bindgen]
+pub fn renumber_phrase(song: &mut WasmSong, phrase: usize, to_phrase: usize) -> Result<bool, String> {
+    if phrase >= Song::N_PHRASES { return Err(format!("Error invalid source phrase number {phrase}")) }
+    if to_phrase >= Song::N_PHRASES { return Err(format!("Error invalid destination phrase number {to_phrase}")) }
+
+    if phrase == to_phrase { return Ok(true) }
+
+    if !song.song.phrases[to_phrase].is_empty() {
+        return Err(format!("Destination phrase {to_phrase} is not empty"))
+    }
+
+    let mut remapper = Remapper::default();
+    remapper.phrase_mapping.mapping[phrase] = to_phrase as u8;
+    remapper.instrument_mapping.to_move.push(phrase as u8);
+    remapper.renumber(&mut song.song);
 
     Ok(true)
 }