+++ src/components/chain_viewer.tsx
@@ -1,0 +1,53 @@
+import * as W from '../../m8-files/pkg/m8_files';
+import { SongPane } from "../state";
+import { hexStr } from "./common";
+
+export function ChainViewer(props: { panel: SongPane }) {
+  const song = props.panel.song.value;
+  const bump = props.panel.bumper.value;
+
+  if (song === undefined) return <div class="rootcolumn"></div>;
+
+  const chain = props.panel.selected_chain.value;
+  if (chain === undefined) {
+    return <div class="rootcolumn">
+      <h4>Chain viewer</h4>
+      <p>Click a chain to view</p>
+    </div>;
+  }
+
+  const chainSteps = W.get_chain_steps(song, chain);
+  const elems = [];
+
+  const phraseSet = (i : number) => {
+    props.panel.selected_phrase.value = i;
+  }
+
+  for (let i = 0; i < 32; i += 2) {
+    elems.push(`${(i / 2).toString(16)} : `)
+    const phrase = chainSteps[i];
+
+    if (phrase === 0xFF) {
+      elems.push("--\n")
+    } else {
+      elems.push(<span class="phrase" onClick={_ => phraseSet(phrase)}>{hexStr(phrase)} {hexStr(chainSteps[i + 1])}</span>)
+      elems.push('\n')
+    }
+  }
+
+  const phrase_idx = props.panel.selected_phrase.value;
+  const phrase = (phrase_idx !== undefined)
+    ? <>
+        <h4>Phrase {phrase_idx}</h4>
+        <pre>{W.show_phrase(song, phrase_idx)}</pre>
+      </>
+    : undefined
+
+  return <div class="chain_viewer">
+    <h4>Chain viewer</h4>
+    <pre>
+      {elems}
+    </pre>
+    {phrase}
+  </div>;
+}