+++ src/state.ts
@@ -1,5 +1,6 @@
 import { Signal, signal } from "@preact/signals";
 import { WasmSong } from "../m8-files/pkg/m8_files";
+import { createContext } from "preact";
 
 export type Position =
     {
@@ -46,20 +47,78 @@ export type ChainNumberEdition =
         current_value: number
     }
 
-export const EmptyEdition : ChainNumberEdition =
+export const EmptyChainEdition : ChainNumberEdition =
     {
         x: -1, y: -1, base_chain: -1, current_value: -1
     };
 
+export type PhraseNumberEdition =
+    {
+        chain_source: number,
+        chain_offset: number,
+        base_phrase: number,
+        current_value: number
+    }
+
+export const EmptyPhraseEdition : PhraseNumberEdition =
+    {
+        chain_source: -1,
+        chain_offset: -1,
+        base_phrase: -1,
+        current_value: -1
+    }
+
+export type InstrumentNameEditor =
+    {
+        instrument_id: number
+        name: string
+    }
+
+export const EmptyInstrumentNameEdition =
+    { instrument_id: -1, name: '' }
+
+export type InstrumentNumberEdition =
+    {
+        base_instrument: number,
+        current_value: number
+    }
+
+export const EmptyInstrumentNumberEidition : InstrumentNumberEdition =
+    { base_instrument: -1, current_value: -1 }
+
 export type SongPane =
     {
-        loaded_name: Signal<string | undefined>,
+        /** Currently loaded song, fully parsed  */
         song: Signal<WasmSong | undefined>,
+
+        /** Loaded file name */
+        loaded_name: Signal<string | undefined>,
+
+        /** Bytes of the raw song, used for rewriting. */
         raw_song: Signal<Uint8Array>,
+
+        /** Signal to force refresh */
         bumper: Signal<number>,
-        edited_chain: Signal<ChainNumberEdition | undefined>,
+
+        /** Currently selected chain in the song view */
         selected_chain: Signal<number | undefined>,
+
+        /** Currently selected phrase within the selected chain */
         selected_phrase: Signal<number | undefined>,
+
+        /** Currently edited chain number */
+        edited_chain: Signal<ChainNumberEdition | undefined>,
+
+        /** Currently edited phrase number */
+        edited_phrase: Signal<PhraseNumberEdition | undefined>,
+
+        /** Currently edited phrase number */
+        edited_instrument: Signal<InstrumentNumberEdition | undefined>,
+
+        /** Instrument currently being renamed */
+        edited_instrument_name: Signal<InstrumentNameEditor | undefined>,
+
+        /** Obsolete for now */
         selection_range: Signal<ChainSelection | undefined>
     }
 
@@ -69,6 +128,9 @@ function initPane() : SongPane {
         song: signal(undefined),
         bumper: signal(0),
         edited_chain: signal(undefined),
+        edited_phrase: signal(undefined),
+        edited_instrument: signal(undefined),
+        edited_instrument_name: signal(undefined),
         raw_song: signal(new Uint8Array(0)),
         selected_chain: signal(undefined),
         selected_phrase: signal(undefined),
@@ -93,10 +155,11 @@ export type State =
     }
 
 export function initState() : State {
-
     return {
         message_banner: signal(undefined),
         left: initPane(),
         right: initPane()
     }
 }
+
+export const GlobalState = createContext<State>(undefined);