+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3
+ 
+ - Song explorer view
+ - Instrument number renumbering using the song explorer
+ - Instrument renaming
+ - Phrase renumbering
+
 # 0.2
 
  - Changing allocation policy to try to keep chain & phrase number